third_party/WebKit/LayoutTests/typedcssom/the-stylepropertymap/declared-expected.txt

This is a testharness.js-based test.
FAIL Declared StylePropertyMap only contains properties in the style rule assert_array_equals: property 1, expected "transition-duration" but got "width"
PASS Declared StylePropertyMap contains CSS property declarations in style rules
PASS Declared StylePropertyMap does not contain inline styles
FAIL Declared StylePropertyMap contains custom property declarations assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
PASS Declared StylePropertyMap does not contain properties with invalid values
PASS Declared StylePropertyMap contains properties with their last valid value
PASS Declared StylePropertyMap is live
Harness: the test ran to completion.
